Refactor CrossPostPage to use CrossPostClient component and move related logic
